refactored startServer.js, server.js into server.js, app.js added forceExit flag to testing
