+++ package.json
@@ -5,11 +5,10 @@
   "main": "server.js",
   "scripts": {
     "frontend": "vite client",
-    "backend": "nodemon server/startServer.js",
-    "client": "npm run dev --prefix client",
+    "backend": "nodemon server/server.js",
     "build": "vite build",
     "preview": "vite preview",
-    "test": "jest --verbose"
+    "test": "jest --runInBand --forceExit --verbose"
   },
   "author": "Goblin Sharks",
   "license": "MIT",
