--- server/startServer.js
@@ -1,9 +1,0 @@
-const app = require('./server.js');
-
-const PORT = process.env.PORT || 8080;
-/**
- * start server
- */
-app.listen(PORT, () => {
-  console.log(`Server listening on port: ${PORT}...`);
-});
